i just built this simple jame for the rock,paper scissors game for the odin project.
i did not follow the instanction but built this
on my own

i was suposed to be played on only the console
but i added a user interface and it could be played

on the web.

it my first time building a logic or somthing like
this using javascript.

thank you for reading , did not do any changes